wifi: ath12k: Add debugfs support to configure primary link
Add a vif specific debugfs entry to select the primary link to be used by
the user in MLO case for the AP mode.

This will helpful while running the functionality and performance test
case.

echo N > /sys/kernel/debug/ieee80211/phy1/netdev:wlan0/primary_link

where N = 0 to set link-0 as primary link, 1 to set link-1 as primary
link and 2 to set link-2 as primary link.

cat /sys/kernel/debug/ieee80211/phy1/netdev:wlan0/primary_link to
get the configured primary link.

Co-developed-by: Avula Sri Charan <asrichar@qti.qualcomm.com>

+++ backports-6.1-9a0dddfb3/include/net/mac80211.h
@@ -2755,7 +2755,7 @@ struct ieee80211_sta {
 
 	struct ieee80211_txq *txq[IEEE80211_NUM_TIDS + 1];
 
-	u16 valid_links;
+	unsigned long valid_links;
 	struct ieee80211_link_sta deflink;
 	struct ieee80211_link_sta __rcu *link[IEEE80211_MLD_MAX_NUM_LINKS];
 	/* multi-link reconfiguration */
